+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 ## 0.9.2 - 2025.xx.xx
 
+- 2025.06.30:
+  - Time stats for migration of tables and data added into all connectors, results are also stored in the protocol tables
+    - Rationale: This will allow to better understand performance of the migration process and identify bottlenecks or situations when migration literally hangs because of some issues on the source database
+    - Migrator now stores detailed time stats for each batch in the new protocol table, and reports basic batch stats at the end of the migration, shortest, longest and average batch time is stored and reported for every table
+  - Adjustments in running parallel workers for migration of tables, data, indexes and constraints to avoid delays in starting the next worker after some of the previous ones finished
+
 ## 0.9.1 - 2025.06.24
 
 - 2025.06.24:
+
   - Add project logo and architecture diagram to PyPI page (@mbanck)
 
 - 2025.06.19: